+++ R/read.NanoSurf_file.R
@@ -276,7 +276,7 @@ read.NID_header <- function(filename) {
   i=0
   dlen.header = 0
   while( TRUE ) {
-    rline = readLines(con,n=1, skipNul=TRUE)
+    rline = readLines(con, n=1, skipNul=TRUE, warn = FALSE)
     if (substr(rline,1,2) == "#!" ) break
     i = i + 1
     dlen.header = dlen.header + nchar(rline, type="bytes") + 2